requirements.txt
fastapi
uvicorn
asyncpg
pydantic-settings
python-dotenv
google-generativeai